backend/requirements.txt
Django==5.1.15
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.29.0
django-cors-headers==4.9.0
django-filter==24.3
Pillow==11.2.0
python-decouple==3.8
psycopg2-binary==2.9.10
